+++ NEWS.md
@@ -1,3 +1,6 @@
+* Update UI to show Admin tab if pertinent.
+* Defined `retrieve_opinions` to retrieve all the last elicitated
+  values each expert. The function can be executed buy admiin only.
 * Move testing computation inside tests to permit to skip them
 * reshaping the app in multiple files
 * Save/restore session implemented